chore: Move AddDataFields folder to new location
